+++ HISTORY.md
@@ -5,6 +5,8 @@
 ## 2.0.0.e (2011-??-??)
 
 * Adding enqueue_at_with_queue/enqueue_in_with_queue support (niralisse)
+* Adding `Resque::Scheduler.poll_sleep_amount` to allow for configuring
+  the sleep time b/w delayed queue polls.
 
 ## 2.0.0.d (2011-04-04)